Update to REP-0003 for ROS Kinetic Kame
I've updated the requirements for Kinetic Kame. @scpeters has expressed interest in ROS having support for C++11 for better compatibility with newer versions of SDFormat. I don't know if the version of Gazebo should be 5 or 6, and the list of C++11 features will be expanded as we see if they can benefit our users (e.g. lambdas for subscriptions, std::shared_ptr instead of boost's)